Build against the leios-prototype dependency pins
Pin cardano-api to its leios-prototype branch and the rest of the
Leios fleet (cardano-base, cardano-ledger, ouroboros-consensus,
ouroboros-network, typed-protocols, kes-agent, plutus) as
source-repository-packages, with the constraints, allow-newer entries
and deprecation-warning relaxations the pinned closure needs, and
adapt the tree to it: the pinned api and ledger add a BLS key field to
stake pool parameters, so the construction sites set it to Nothing;
the pinned api keeps the DRep-deposit arguments of the experimental
balancing functions, so their threading returns; the pinned network
fork predates cardano-diffusion 1.1, so the ping command and its error
type return to their pre-1.1 implementations and the relay
ping-validation feature is carried out of the branch until the fork
rebases; the pinned ledger's minor protocol version is Natural,
bridged with fromIntegral; and the Windows CI job enables
core.longpaths for the deep SRP checkouts.

+++ cardano-cli/src/Cardano/CLI/EraBased/Governance/Actions/Option.hs
@@ -237,7 +237,7 @@ toNonNegativeIntervalOrErr r = case L.boundRational r of
 
 mkProtocolVersionOrErr :: (Natural, Word32) -> L.ProtVer
 mkProtocolVersionOrErr (majorProtVer, minorProtVer) =
-  case (`L.ProtVer` minorProtVer) <$> L.mkVersion majorProtVer of
+  case (`L.ProtVer` fromIntegral minorProtVer) <$> L.mkVersion majorProtVer of
     Just v -> v
     Nothing ->
       error $ "mkProtocolVersionOrErr: invalid protocol version " <> show (majorProtVer, minorProtVer)
